Move DB filters to their own file; enable 2 feature flags in local cfg
